refactor: relocate auth server actions to lib/actions
Move src/lib/auth/actions.ts to src/lib/actions/auth-actions.ts to
align with the project convention that all Server Actions live under
lib/actions/. Updates the three component import sites accordingly.
No behavior change.
