+++ src/components/layout/user-menu-presenter.tsx
@@ -10,7 +10,7 @@ import {
   DropdownMenuTrigger,
 } from '@/components/ui/dropdown-menu';
 import { migrateGuestToGoogleAction } from '@/lib/actions/guest-actions';
-import { signOutAction } from '@/lib/auth/actions';
+import { signOutAction } from '@/lib/actions/auth-actions';
 import Link from 'next/link';
 import { GuestSignOutDialog } from './guest-sign-out-dialog';
